Refactor auth logic into component
